// Features/reports/Types/IExtraHourForReport.ts
export interface IExtraHourForReport {
    identifier: string;
    idExtraHourLateness: number;
    idEmployee: number;
    fullName: string;
    costCenter: string;
    salary: number;
    percentValue: number;
    concept: string;
    hourAmount: number;
    amount: number;
    date: string;
    isPaid: string;
    payrollName: string;
    idPayrollPay: number;
    idPosition: number;
    idDepartment: number;
    position: string;
    department: string;
    idCostCenter: number;
    accountNumber: string;
    conceptCode: string;
    idCompany: number;
}